The remove button for active users on both row and square cards effectively deletes the employee.

+++ backend/app/authentication/repository.py
@@ -31,15 +31,29 @@ def get_pending_user_by_id(db: Session, id: int):
 def get_pending_user_by_email(db: Session, email: str):
     return db.query(PendingUser).filter(PendingUser.email == email).first()
 
-def delete_pending_user(db: Session, user: PendingUser):
-    db.delete(user)
-    db.commit()
+def delete_pending_user(db: Session, user_id: int):
+    user = db.query(PendingUser).filter(PendingUser.user_id == user_id).first()
+    
+    if user:
+        db.delete(user)
+        db.commit()
+    
+    return user
 
 def create_user(db: Session, firstname: str, surname: str, username: str, email: str, oid: str, role: str, driveId: str, refresh: bytes) -> User:
     user = User(firstname=firstname, surname=surname, username=username, email=email, oid=oid, role=role, driveId=driveId, refresh=refresh)
     db.add(user)
     db.commit()
     db.refresh(user)
+    return user
+
+def delete_user(db: Session, user_id: int):
+    user = db.query(User).filter(User.user_id == user_id).first()
+    
+    if user:
+        db.delete(user)
+        db.commit()
+    
     return user
 
 def verify_refresh(hashed_token: str, expiry: datetime, db: Session) -> Refresh:
